pyproject.toml
[tool.poetry]
name = "trading-playground"
version = "0.1.0"
description = "A learning-focused trading system for stocks and cryptocurrency markets"
authors = ["Whit Walters <whitney.walters@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.0"
numpy = "^1.26.0"
scipy = "^1.12.0"
scikit-learn = "^1.4.0"
pytz = "*"
requests = "*"
aiohttp = "*"
websockets = "*"
pydantic = "*"
pydantic-settings = "*"
asyncio = "*"
plotly = "*"
pyyaml = "^6.0.2"
python-dotenv = "^1.0.1"
yfinance = "^0.2.36"
polygon-api-client = "^1.12.3"
asyncpg = "^0.29.0"
pytest-asyncio = "^0.25.2"
statsmodels = "0.14.1"

# Django dependencies
django = "^5.0.3"
djangorestframework = "^3.14.0"
django-filter = "^23.5"
django-cors-headers = "^4.3.1"
djangorestframework-simplejwt = "^5.3.1"
channels = "^4.0.0"
channels-redis = "^4.1.0"
daphne = "^4.0.0"
celery = "^5.3.6"
redis = "^5.0.1"
django-redis = "^5.4.0"
psycopg = "^3.1.18"
uvicorn = "^0.27.1"

[tool.poetry.group.dev.dependencies]
pytest = "*"
pytest-asyncio = "*"
black = "*"
mypy = "*"
pytest-cov = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
